requirements.txt
fastapi>=0.115,<1.0
minio>=7.2,<8.0
python-dotenv>=1.0,<2.0
python-multipart>=0.0.9,<1.0
sqlmodel>=0.0.22,<1.0
uvicorn[standard]>=0.32,<1.0
